cookie consent

+++ pages/api/users/[userId].js
@@ -6,26 +6,31 @@ import {
 } from '../../../util/database';
 
 export default async function handler(req, res) {
-  if (req.method === 'GET') {
-    // console.log('from API query', req.query.userId);
-    const token = req.cookies.sessionToken;
-    const session = await getValidSessionByToken(token);
+  const token = req.cookies.sessionToken;
+  const session = await getValidSessionByToken(token);
 
-    if (!session) {
-      res.status(404).send({
-        errors: [{ message: 'Not a valid Session' }],
-      });
-      return;
-    }
+  console.log('req', req.query.userId);
+
+  console.log('session', session.userId);
+
+  if (!session) {
+    res.status(404).send({
+      errors: [{ message: 'Not a valid Session' }],
+    });
+    return;
+  } else if (
+    req.method === 'GET' &&
+    session.userId === Number(req.query.userId)
+  ) {
     const user = await getUser(Number(req.query.userId));
     res.status(200).json(user);
-  } else if (req.method === 'DELETE') {
+  } else if (
+    req.method === 'DELETE' &&
+    session.userId === Number(req.query.userId)
+  ) {
     const deletedUser = await deleteUserByUsername(req.query.userId);
-
-    // console.log('from API', deleteUserByUsername());
     return res.status(200).json(deletedUser);
   }
-
   return res.status(405).end();
 }
